Move global vars changing tests into run-pass
Should fix race #42795
